Implement an anonymous composite cache.
By using an explicit cache for anonymous composites, it allows re-use of blessed
TupleDesc's. This is important to mitigate leaks in certain
situations such as table returning functions and now Postgres.Statement's.

In statement's case, fully anonymous TupleDesc's were being used. This makes
working with the resulting DatumTuples rather difficult as the standard
operators cannot be used with records that do reference a blessed TupleDesc.

The misfortune here is that Postgres.Statement will never release the `output`
type when the statement returns rows.

fixes #5